+++ src/activities/util/IntervalSelectionActivity.cpp
@@ -63,7 +63,7 @@ int IntervalSelectionActivity::clampedValue(const int candidate) const {
 
 bool IntervalSelectionActivity::usesTextTouchStepControls() const {
   return titleId == StrId::STR_TIME_TO_SLEEP || titleId == StrId::STR_AUTO_TURN_INTERVAL_SECONDS ||
-         titleId == StrId::STR_LINE_SPACING;
+         titleId == StrId::STR_LINE_SPACING || titleId == StrId::STR_TOP_BOTTOM || titleId == StrId::STR_LEFT_RIGHT;
 }
 
 void IntervalSelectionActivity::onEnter() {